requirements.txt
streamlit>=1.36.0
pandas>=2.2.0
openpyxl>=3.1.2
pillow>=10.0.0
pytesseract>=0.3.10
rapidfuzz>=3.6.1
Unidecode>=1.3.8
opencv-python-headless>=4.10.0.84
gTTS>=2.5.1
easyocr>=1.7.2
tesseract-ocr>=0.1.9